requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
python-dotenv>=1.0
Pillow>=10.0


httpx==0.27.0          
celery==5.3.6         
redis==5.0.3           # Celery broker
djangorestframework>=3.14.0
requests>=2.31.0
